+++ packages/bridge-status-controller/src/bridge-status-controller.batch-sell.test.ts
@@ -165,8 +165,9 @@ const mockQuotes = mockBatchSellErc20Erc20
       gasSponsored: undefined,
     },
   }))
-  .map((quote) =>
-    mergeQuoteMetadata(quote, {
+  .map((quote) => ({
+    ...quote,
+    ...{
       sentAmount: {
         usd: '100',
         valueInCurrency: '200',
@@ -175,8 +176,8 @@ const mockQuotes = mockBatchSellErc20Erc20
         usd: '101',
         valueInCurrency: '201',
       },
-    }),
-  );
+    },
+  }));
 const mockTransferTx: BatchSellTradesResponse['transactions'][number] = {
   chainId: 10,
   from: '0xaccount1',